+++ src/components/main/tab/DiagramsTabPane.vue
@@ -49,7 +49,7 @@ import { Component, Vue, Emit, Prop } from 'nuxt-property-decorator'
 import DiagramEditorContainer from '@/components/diagrams/DiagramEditorContainer.vue'
 import Diagram from '@/domain/diagram/Diagram'
 import Resource from '@/domain/resource/Resource'
-import TreeItem from '@/presentation/tree/TreeItem'
+import TreeItem from '@/components/main/tree/TreeItem'
 
 @Component({
   components: {
