+++ src/components/main/tree/FolderTreeFactory.ts
@@ -1,5 +1,5 @@
 import Folder from "./Folder";
-import TreeItem from "@/presentation/tree/TreeItem";
+import TreeItem from "@/components/main/tree/TreeItem";
 import DiagramType from "@/domain/diagram/DiagramType";
 import Diagrams from "@/domain/diagram/Diagrams";
 import Diagram from "@/domain/diagram/Diagram";
